Add delete item permissions check

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -47,9 +47,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title}`);
-    // 2. check if they won the item, or have permissions
-    // TODO
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+    // 2. check if they own the item, or have permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -176,13 +183,13 @@ const Mutations = {
   },
   async updatePermissions(parent, args, ctx, info) {
     // 1. Check if they are logged in
-    // if (!ctx.request.userId) {
-    //   throw new Error("You must be logged in!");
-    // }
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in!");
+    }
 
     // 2. Query the current user
     const currentUser = await ctx.db.query.user(
-      { where: { id: args.userId } },
+      { where: { id: ctx.request.userId } },
       info
     );
 
@@ -190,7 +197,6 @@ const Mutations = {
     hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
 
     // 4. Update the permissions
-
     return ctx.db.mutation.updateUser(
       {
         data: {
